migrations/models/3_20240916120715_added_purchased_price_column.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `operations` ADD `purchased_price` DOUBLE;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `operations` DROP COLUMN `purchased_price`;"""
