requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dictdiffer==0.9.0
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
html5lib==1.1
idna==3.6
iso8601==1.1.0
kiwisolver==1.4.5
lxml==5.2.2
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.0.5
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.1
pyparsing==3.1.4
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
ta==0.11.0
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.41
